requirements.txt
bettercam>=1.6.9.8.4
mss>=9.0.1
winrt-runtime>=2.0.0
winrt-Windows.Media.Ocr>=2.0.0
winrt-Windows.Graphics.Imaging>=2.0.0
winrt-Windows.Storage.Streams>=2.0.0
winrt-Windows.Globalization>=2.0.0
winrt-Windows.Foundation>=2.0.0
winrt-Windows.Foundation.Collections>=2.0.0
torch>=2.0.0
transformers>=4.40.0
sentencepiece>=0.2.0
PyQt5>=5.15.0
opencv-python>=4.8.0
pynput>=1.7.6
pywin32>=306
pyyaml>=6.0
numpy>=1.24.0
requests>=2.31.0
huggingface-hub>=0.20.0
